+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## Upcoming
+
+* Allow `NULL` values for:
+  `TextDigestField`;
+  `TextHMACField`;
+  `IntegerPGPPublicKeyField`;
+  `IntegerPGPSymmetricKeyField`.
+
 ## v0.6.1
 
 * Fix `cast`ing bug when sending negative values to integer fields.